change: moves circuit breaker tests to existing resilience decorator test file
