+++ test/decorators/use-resilience.decorator.spec.ts
@@ -1,12 +1,16 @@
-import { FixedBackoff, RetryStrategy, TimeoutStrategy, UseResilience } from '../../src';
+import {
+	CircuitBreakerStrategy,
+	FixedBackoff,
+	RetryStrategy,
+	TimeoutStrategy,
+	UseResilience
+} from '../../src';
 
 const timeoutStrategy = new TimeoutStrategy(100);
 const retryStrategy = new RetryStrategy({
 	maxRetries: 3,
 	backoff: FixedBackoff
 });
-
-const spyOnRetry = jest.spyOn(FixedBackoff.prototype, 'getGenerator');
 
 class UserService {
 	@UseResilience(timeoutStrategy)
@@ -33,10 +37,29 @@ class UserService {
 
 		return [{ id: '1', name: 'John Doe' }];
 	}
+
+	async queryUser(id: string) {
+		return { id, name: 'John Doe' };
+	}
+
+	@UseResilience(
+		new CircuitBreakerStrategy({
+			requestVolumeThreshold: 3,
+			rollingWindowInMilliseconds: 1000
+		})
+	)
+	async getUserCircuitStrategy(id: string) {
+		return await this.queryUser(id);
+	}
 }
 
 describe('Resilience Decorator', () => {
 	const userService = new UserService();
+
+	beforeEach(async () => {
+		await jest.advanceTimersByTimeAsync(10000);
+		jest.restoreAllMocks();
+	});
 
 	it('should be able to execute a command', async () => {
 		const value = await userService.getUser('1');
@@ -52,5 +75,25 @@ describe('Resilience Decorator', () => {
 			expect(value).toEqual([{ id: '1', name: 'John Doe' }]);
 			done();
 		});
+	});
+
+	it('should open the circuit and throw a circuit opened error', async () => {
+		jest.spyOn(userService, 'queryUser').mockRejectedValue(new Error('err'));
+
+		// Execute multiple calls to trigger circuit opening
+		await expect(async () => await userService.getUserCircuitStrategy('1')).rejects.toThrow(
+			new Error('err')
+		); // 1st call. 20% failures
+		await expect(async () => await userService.getUserCircuitStrategy('1')).rejects.toThrow(
+			new Error('err')
+		); // 2nd call. 40% failures
+		await expect(async () => await userService.getUserCircuitStrategy('1')).rejects.toThrow(
+			new Error('err')
+		); // 3rd call. 60% failures
+
+		// Circuit should be open now, expect circuit opened error
+		await expect(async () => await userService.getUserCircuitStrategy('1')).rejects.toThrow(
+			new Error('Circuit is open')
+		);
 	});
 });
